move some more utils, add city guide folder
