+++ package.json
@@ -31,15 +31,15 @@
     "clean-ios": "yarn rimraf ~/Library/Developer/Xcode/DerivedData",
     "codepush:deploy": "./scripts/codepush/deploy-to-codepush.sh",
     "danger": "danger",
-    "dead-code": "./scripts/check-dead-code.sh",
+    "dead-code": "./scripts/utils/check-dead-code.sh",
     "deploy-beta": "./scripts/deploy-beta-both",
     "detect-secrets:hook": "scripts/detect-secrets/detect-secrets.sh hook",
     "detect-secrets:rebuild": "scripts/detect-secrets/detect-secrets.sh rebuild",
-    "doctor": "./scripts/doctor.js",
-    "flip_table": "./scripts/flip-table",
-    "flip_table_extreme": "./scripts/flip-table-extreme",
-    "generate-cities-cache": "node scripts/generate-cities-cache.js",
-    "generate-cities-objc": "node scripts/generate-cities-objc.js",
+    "doctor": "./scripts/utils/doctor.js",
+    "flip_table": "./scripts/utils/flip-table",
+    "flip_table_extreme": "./scripts/utils/flip-table-extreme",
+    "generate-cities-cache": "node scripts/city-guide/generate-cities-cache.js",
+    "generate-cities-objc": "node scripts/city-guide/generate-cities-objc.js",
     "generate-changelog": "ts-node scripts/changelog/generateChangelog.ts",
     "navigate-to-route": "./scripts/navigate-to-route",
     "navigate-to-route-android": "./scripts/navigate-to-route-android",
