+++ scripts/utils/doctor.js
